% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_get_ftpal.R
\name{barcolor}
\alias{barcolor}
\title{Get bar color}
\usage{
barcolor(index = NA, type = NA, take = NA)
}
\arguments{
\item{index}{Index (row number) of palette}

\item{type}{F for first take, H for home take}

\item{take}{Integer, which video to use}
}
\description{
Get bar color
}
