% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_palette.R
\name{create_palette}
\alias{create_palette}
\title{Create palette}
\usage{
create_palette(video_id, n = 6, maxiter = 45, monitor = plot, parallel = F)
}
\arguments{
\item{video_id}{youtube video ID}

\item{n}{number of colors to create (default 6)}

\item{maxiter}{Max number of iterations for the genetic algorithm}

\item{monitor}{set to plot to view convergence in progress}

\item{parallel}{number of cores to use for parallel processing (F by default)}
}
\value{
A color palette, in hex strings
}
\description{
Create a palette from the thumbnail of the given video. Requires the GA
package. Parallel processing possible but turned off by default. This actually
works with any thumbnail, doesn't have to be from the first take.
}
