% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA_binary_to_rgb.R
\name{.binary_to_rgb}
\alias{.binary_to_rgb}
\title{Binary chromosome to RGB representation}
\usage{
.binary_to_rgb(binary_chromosome)
}
\arguments{
\item{binary_chromosome}{Binary representation of chromosome}
}
\value{
RGB representation of chromosome
}
\description{
Binary chromosome to RGB representation
}
