% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_utils.R
\name{.calculate_backtrack}
\alias{.calculate_backtrack}
\title{Backtracking through playlists}
\usage{
.calculate_backtrack(playlists)
}
\arguments{
\item{playlists}{Playlist dataframe from .get_playlists()}
}
\description{
Backtracking through playlists
}
