% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA_utils.R
\name{.chromosome_as_binary}
\alias{.chromosome_as_binary}
\title{RGB chromosome to binary representation}
\usage{
.chromosome_as_binary(chromosome)
}
\arguments{
\item{chromosome}{RGB value representation of the chromosome}
}
\value{
Create a binary representation of the palette chromosome
}
\description{
RGB chromosome to binary representation
}
