% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA_utils.R
\name{.color_fitness}
\alias{.color_fitness}
\title{Fitness function}
\usage{
.color_fitness(new_chromosome, old_chromosome, contiguity_matrix, pal_size = 6)
}
\arguments{
\item{new_chromosome}{New binary chromosome}

\item{old_chromosome}{Original binary chromosome from kmeans palette}

\item{contiguity_matrix}{Matrix for comparisons for cj calculation}

\item{pal_size}{Size of the palette}
}
\value{
A fitness score
}
\description{
Equation 7, used in the genetic algorithm
}
