% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA_contrast_ratio.R
\name{.contrast_ratio}
\alias{.contrast_ratio}
\title{Compute contrast ratio}
\usage{
.contrast_ratio(color1, color2)
}
\arguments{
\item{color1}{First color, in RGB vector format}

\item{color2}{Second color, in RGB vector format}
}
\value{
C, the contrast ratio
}
\description{
Equation 4
}
