% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_convert_deut.R
\name{.convert_deut}
\alias{.convert_deut}
\title{Convert to deuteronopia color space}
\usage{
.convert_deut(hex_palette)
}
\arguments{
\item{hex_palette}{Hex strings of a palette}
}
\value{
New palette with new colors
}
\description{
Helper to quickly convert to deuteronopia color vision
}
