% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA_fit_cj.R
\name{.fit_cj}
\alias{.fit_cj}
\title{Fit contrast ratios}
\usage{
.fit_cj(new_mat, contiguity_matrix, threshold = 5)
}
\arguments{
\item{new_mat}{Matrix of rgb colors}

\item{contiguity_matrix}{Matrix for comparisons}

\item{threshold}{Fit threshold}
}
\value{
Vector of cj values
}
\description{
Equation 9. There is a slight modification where I include pure black
and pure white as additional comparisons to avoid the algorithm settling on
those values. This also has an effect of reducing saturation more generally.
}
\details{
The threshold can be modified, especially for the black/white comparisons,
but I had to set them >6 to get interesting results. You could try using
different values for each one.
}
