% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA_fit_di.R
\name{.fit_di}
\alias{.fit_di}
\title{Color distance}
\usage{
.fit_di(new_mat, old_mat)
}
\arguments{
\item{new_mat}{Matrix of rgb colors}

\item{old_mat}{Original matrix of rgb colors}
}
\value{
Vector of distance values to use in the fitness function
}
\description{
Equation 8
}
