% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_generate_all_palettes.R
\name{.generate_all_palettes}
\alias{.generate_all_palettes}
\title{Generate all palettes}
\usage{
.generate_all_palettes(first_takes = ftpals::first_takes)
}
\arguments{
\item{first_takes}{Palette table}
}
\description{
Creates creates the palettes in bulk. Takes about an hour on my computer for
117 videos.
}
