% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA_generate_comparisons.R
\name{.generate_comparisons}
\alias{.generate_comparisons}
\title{Generate comparisons}
\usage{
.generate_comparisons(contiguity_matrix)
}
\arguments{
\item{contiguity_matrix}{Binary matrix for comparisons}
}
\description{
Returns indices to use in comparisons
}
