% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA_utils.R
\name{.generate_matrix}
\alias{.generate_matrix}
\title{Generate comparison matrix.}
\usage{
.generate_matrix(n = 6)
}
\arguments{
\item{n}{Size of the palette}
}
\value{
A comparison matrix
}
\description{
The comparisons I decided on is such that each color is compared to the
two colors it's next two, as well as every other color in the palette.
So, all the odd values are compared, and all the even values are compared.
}
\details{
This matrix is a slight modification on the matrices (models) used in
Troiano, Birtolo, and Miranda 2008
}
