% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_get_all_videos.R
\name{.get_all_videos}
\alias{.get_all_videos}
\title{Get all videos}
\usage{
.get_all_videos()
}
\value{
Data frame with all the first take videos
}
\description{
Procedure to scrape all the information about the videos
}
