% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA_utils.R
\name{.get_reduced_components}
\alias{.get_reduced_components}
\title{Reduce to LMS color space}
\usage{
.get_reduced_components(rgb_color)
}
\arguments{
\item{rgb_color}{RGB vector of color}
}
\description{
Equation 10, 12, 13
}
