% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_utils.R
\name{.get_take_number}
\alias{.get_take_number}
\title{Get take number from video id}
\usage{
.get_take_number(video_id)
}
\arguments{
\item{video_id}{First take video ID to look up}
}
\value{
The number for the video
}
\description{
Get take number from video id
}
