% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_utils.R
\name{.get_videos}
\alias{.get_videos}
\title{Get videos}
\usage{
.get_videos(
 
    playlist_url = "https://www.youtube.com/playlist?list=PLeLvSt3A0Ddk9lGjNqmzT0ctVPA5xfinn"
)
}
\arguments{
\item{playlist_url}{Set to the "THE FIRST TAKE" playlist by default}
}
\description{
Retrieve information about videos on a page, works with the video
upload page or with a playlist page.
}
