% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA_lab_edist.R
\name{.lab_edist}
\alias{.lab_edist}
\title{Compute LAB color space euclidean distance}
\usage{
.lab_edist(color1 = c(0, 255, 0), color2 = c(0, 0, 255))
}
\arguments{
\item{color1}{Color, RGB vector format. Default green.}

\item{color2}{Color, RGB vector format. Default Blue}
}
\value{
Euclidean distance between two colors in LAB colorspace
}
\description{
Equation 3. Max distance between Green and Blue
}
