% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_manual_fix_errors.R
\name{.manual_fix_errors}
\alias{.manual_fix_errors}
\title{Fix errors}
\usage{
.manual_fix_errors(first_takes)
}
\arguments{
\item{first_takes}{Palette table}
}
\value{
Cleaned up first take table
}
\description{
Some videos just have random typos and errors that have to be corrected. Note
that there is NO first take #70
}
