% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA_utils.R
\name{.optimize_palette}
\alias{.optimize_palette}
\title{Optimize palette}
\usage{
.optimize_palette(palette, maxiter = 45, parallel = F, monitor = plot)
}
\arguments{
\item{palette}{Original palette, in hex strings}

\item{maxiter}{Number of iterations to run}

\item{parallel}{Should you run in parallel? F to turn off, else # of cores to use}

\item{monitor}{Monitor convergence? use plot to view, F to turn off}
}
\value{
A hex string with a new palette
}
\description{
Given a palete, run the genetic algorithm described in Troiano, Birtolo,
and Miranda 2008 to create a palette that should work better for those
with colorblindness.
}
