% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_order_palette.R
\name{.order_palette}
\alias{.order_palette}
\title{Order Palette}
\usage{
.order_palette(hex_palette)
}
\arguments{
\item{hex_palette}{Palette, given in hex strings}
}
\value{
A palette ordered by the given procedure
}
\description{
The ordering these palettes use is like so:
-Order by increasing luminance
-Reorder by flip flopping between high and low values
}
