% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA_palette_to_chromosome.R
\name{.palette_to_chromosome}
\alias{.palette_to_chromosome}
\title{Get chromosome of palette}
\usage{
.palette_to_chromosome(hex_palette)
}
\arguments{
\item{hex_palette}{Hex strings for palette}
}
\value{
a vector of RGB values
}
\description{
Get chromosome of palette
}
