% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA_relative_luminance.R
\name{.relative_luminance}
\alias{.relative_luminance}
\title{Relative Luminance}
\usage{
.relative_luminance(color)
}
\arguments{
\item{color}{Color, in RGB vector format}
}
\value{
L, the relative luminance
}
\description{
Equation 5
}
