% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_retrieve_playlists.R
\name{.retrieve_playlists}
\alias{.retrieve_playlists}
\title{Get playlists from First Take channel}
\usage{
.retrieve_playlists()
}
\value{
Playlists from channel
}
\description{
Get playlists from First Take channel
}
