% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_rgb_to_hex.R
\name{.rgb_to_hex}
\alias{.rgb_to_hex}
\title{RGB values to hex palette}
\usage{
.rgb_to_hex(rgb_vector)
}
\arguments{
\item{rgb_vector}{RGB vector for a color}
}
\description{
Convert from an rgb representation to hex representation
}
