% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA_RGB_to_rgb.R
\name{.RGB_to_rgb}
\alias{.RGB_to_rgb}
\title{RGB to rgb conversion}
\usage{
.RGB_to_rgb(color)
}
\arguments{
\item{color}{RGB vector of color}
}
\value{
Returns a vector of small rgb values
}
\description{
Equation 6
}
