% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_scale_color.R
\name{.scale_color}
\alias{.scale_color}
\title{Scale color}
\usage{
.scale_color(color)
}
\arguments{
\item{color}{Color, in RGB vector format}
}
\value{
Vector of doubles in [0,1]
}
\description{
Just a helper to scale colors to [0,1]
}
