% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_utils.R
\name{.scrape_take_numbers}
\alias{.scrape_take_numbers}
\title{Scrape take numbers in bulk}
\usage{
.scrape_take_numbers(video_df)
}
\arguments{
\item{video_df}{Video data frame from get_videos()}
}
\value{
Vector of take numbers
}
\description{
Scrape take numbers in bulk
}
