% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_take.R
\name{first_take}
\alias{first_take}
\title{Listen to a random song}
\usage{
first_take(
  index = sample(seq_len(nrow(takes)), 1),
  takes = ftpals::first_takes
)
}
\arguments{
\item{index}{Which song to listen to (index in the palette table)}

\item{takes}{Palette table}
}
\description{
Run to open a random song in youtube. Provide an index to listen to a specific
song.
}
