% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_get_ftpal.R
\name{get_ftpal}
\alias{get_ftpal}
\title{Get a palette}
\usage{
get_ftpal(index = NA, palette = 1L, type = NA, take = NA)
}
\arguments{
\item{index}{Index (row number) of palette}

\item{palette}{Palette}

\item{type}{F for first take, H for home take}

\item{take}{Integer, which video to use}
}
\value{
A hex string vector for the palette
}
\description{
Get a palette
}
