% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_thumbnail.R
\name{get_thumbnail}
\alias{get_thumbnail}
\title{Get thumbnail of video}
\usage{
get_thumbnail(index, palettes = ftpals::first_takes)
}
\arguments{
\item{index}{Which video to get?}

\item{palettes}{Palette table}
}
\value{
A raster image
}
\description{
You can use this to add a thumbnail as an annotation to ggplots
}
