% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_plot_ftpalette.R
\name{plot_ftpalette}
\alias{plot_ftpalette}
\title{View palette compared to thumbnail}
\usage{
plot_ftpalette(
  index,
  palettes = ftpals::first_takes,
  use_theme = ggplot2::theme_gray,
  show_thumbnail = TRUE
)
}
\arguments{
\item{index}{Which video to get? an integer}

\item{palettes}{the palette table}

\item{use_theme}{a theme to use, such as theme_minimal, theme_gray, theme_dark, etc.}

\item{show_thumbnail}{Should the thumbnail be downloaded and shown? Requires RCurl}
}
\description{
View palette compared to thumbnail
}
