% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_show_pals.R
\name{show_pals}
\alias{show_pals}
\title{Show colors}
\usage{
show_pals(index = NA, type = "F", take = NA, palettes = ftpals::first_takes)
}
\arguments{
\item{index}{Which palette to get?}

\item{type}{"F" for first take, "H" for home take}

\item{take}{Take #}

\item{palettes}{Palette table}
}
\description{
Helper to visualize both palettes from the palette table with the given index
}
