% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_show_pals.R
\name{show_pals}
\alias{show_pals}
\title{Show colors}
\usage{
show_pals(index, palettes = ftpals::first_takes)
}
\arguments{
\item{index}{Which palette to get?}

\item{palettes}{Palette table}
}
\description{
Helper to visualize both palettes from the palette table with the given index
}
